+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Versions
 
+## 1.2.4
+
+- Retry connecting to Telegram bot every minute if failed on start
+- Display rpc connection error in plain Engligh
+- Add github workflow to build docker image
+
 ## 1.2.3
 
 - Add fee rate estimate for peg-in